refactor: final 키워드 적용 및 domain 패키지 세분화
