+++ src/main/java/blackjack/domain/GameResult.java
@@ -1,5 +1,8 @@
 package blackjack.domain;
 
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
+
 public enum GameResult {
     WIN,
     DRAW,
