+++ src/main/java/blackjack/domain/card/CardDeck.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.Collections;
 import java.util.List;
