+++ src/main/java/blackjack/domain/card/Denomination.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Denomination {
     ACE("A", 1, 11),
